+++ webpack.config.js
@@ -21,7 +21,7 @@ module.exports = {
       {
         test: /\.tsx?$/,
         use: "ts-loader",
-        exclude: /node_modules/,
+        exclude: [/node_modules/, /\.test\.ts$/],
       },
       {
         test: /\.css$/,